class-39-django-rest-framework/demos/kanban-39/web/kanban_rest/serializers.py
from django.contrib.auth.models import User
from ..board.models import Category, Card
from rest_framework import serializers


class UserSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)

    class Meta:
        model = User
        fields = ('id', 'username', 'email', 'password', 'first_name', 'last_name')

    def create(self, validated_data):
        user = super().create({
            'username': validated_data['username'],
            'email': validated_data['email'],
        })

        user.set_password(validated_data['password'])
        user.save()
        return user


class CategorySerializer(serializers.HyperlinkedModelSerializer):
    owner = serializers.ReadOnlyField(source='user.username')
    user = serializers.HyperlinkedRelatedField(view_name='user-detail', read_only=True)

    class Meta:
        model = Category
        fields = ('id', 'owner', 'name', 'description', 'user')


class CardSerializer(serializers.HyperlinkedModelSerializer):
    category = serializers.HyperlinkedRelatedField(view_name='category-api', read_only=True)

    class Meta:
        model = Card
        fields = ('id', 'assigned_to', 'category', 'title', 'description', 'status')

