class-39-django-rest-framework/demos/kanban-39/web/requirements.txt
confusable-homoglyphs==3.2.0
Django==2.1.1
django-cors-headers==2.4.0
django-registration==3.0
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2-binary==2.7.5
pytz==2018.5
